Revert "prove search"
This reverts commit 6b50080586883a52dc7f47a6c06a0a3629d57d94.

+++ app/src/main/java/com/example/readsapp/models/VolumeInfo.java
@@ -9,7 +9,7 @@ public class VolumeInfo{
     private String publisher;
     private String publishedDate;
     private Long pageCount;
-    private int averageRating;
+    private double averageRating;
     private ImageLinks imageLinks;
     private String language;
     private String previewLink;
@@ -63,11 +63,11 @@ public class VolumeInfo{
         this.pageCount = pageCount;
     }
 
-    public int getAverageRating() {
+    public double getAverageRating() {
         return averageRating;
     }
 
-    public void setAverageRating(int averageRating) {
+    public void setAverageRating(double averageRating) {
         this.averageRating = averageRating;
     }
